# src/main/java/com/nttdata/bank/response/BalanceResponse.java
package com.nttdata.bank.response;

import com.fasterxml.jackson.annotation.JsonInclude;
import lombok.Data;

@Data
@JsonInclude(JsonInclude.Include.NON_NULL)
public class BalanceResponse {

	private String accountNumber;
	private Double amount;
	
}